+++ microservice/microservice/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Beam scheduler jobs have a new `SuspendedAt` property
 - Add the ability to use remote storage in local microservices. Can be done by calling `MicroserviceBootstrapper.ForceUseRemoteDependencies<TMicroservice>()`
 in the `Program.cs` right after the call for `await MicroserviceBootstrapper.Prepare<TMicroservice>()`.
+- Callable methods' `Context` field have access to `AccountId`, `GamePid`, and `BeamContext` properties.
 
 ### Changed
 - `BeamScheduler.GetJobs` is obsolete and `GetAllJobs` should be used instead